fix: adopt suggested changes from review
