+++ services/task-svc/internal/room/models/models.go
@@ -1,0 +1,18 @@
+package models
+
+import "github.com/google/uuid"
+
+type Base struct {
+	Name string `gorm:"column:name"`
+}
+
+type Room struct {
+	Base
+	ID   uuid.UUID `gorm:"column:id"`
+	Beds []Bed     `gorm:"foreignKey:RoomID"`
+}
+
+type Bed struct {
+	ID     uuid.UUID `gorm:"column:id"`
+	RoomID uuid.UUID `gorm:"column:room_id"`
+}
